Add support for authenticating
Prior to this commit, Tentabucket could only make calls to API endpoints
that did not require any authentication as a specific user, which was a
quite limited set. This commit adds the ability to specify either
username/password credentials (covering regular username/password
situations and app passwords, particularly for those with 2FA enabled)
and token credentials (covering OAuth situations) and the appropriate
property-based / generative tests to validate functionality across the
valid input spaces for both.

In the unusual case of username/password and token as available
credentials, we prefer the token mechanisms as it's assumed to be, ever
so slightly, more secure than regular or app passwords.